requirements.txt
pdfplumber==0.10.4
python-docx==1.1.0
spacy==3.7.4
sentence-transformers==2.2.2
streamlit==1.32.0
pyyaml==6.0.1
numpy==1.26.4
torch==2.2.1